backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator>=2.0.0

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1
redis>=4.5.0

# AI/ML
torch>=2.0.0
transformers==4.26.0
sentence-transformers==2.2.2
huggingface_hub==0.13.3
tokenizers==0.13.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
matplotlib>=3.7.0
chromadb>=0.4.0
ollama==0.1.6

# 工具
python-dotenv==1.0.0
loguru>=0.7.0
pytest==7.4.3
httpx>=0.25.2
tenacity>=8.2.0
pillow>=10.0.0
langchain==0.0.335
pillow>=10.0.0 